feat: Move parsers into their own translation units to decrease build time
## Build
- Move parsers into their own translation units
- Refactor CMake files for consistency
