+++ cmake/SporeCodegen.cmake
@@ -8,7 +8,7 @@ function(spore_codegen SPORE_TARGET_NAME)
   )
 
   if (NOT SPORE_CODEGEN_BIN_NAME)
-    set(SPORE_CODEGEN_BIN_NAME $<TARGET_FILE:${SPORE_CODEGEN_BINARY_TARGET}>)
+    set(SPORE_CODEGEN_BIN_NAME $<TARGET_FILE:${SPORE_CODEGEN_BIN}>)
   endif ()
 
   if (NOT SPORE_CODEGEN_TARGET_NAME)
