+++ examples/json/CMakeLists.txt
@@ -1,27 +1,17 @@
-set(
-  SPORE_CODEGEN_EXAMPLES_JSON_TARGET "${PROJECT_NAME}-examples-json-bin"
-  CACHE STRING "Target for the JSON serialization example"
-)
+set(TARGET_NAME ${PROJECT_NAME}-examples-json)
 
-file(
-  GLOB_RECURSE SPORE_CODEGEN_EXAMPLES_JSON_FILES
-    ${CMAKE_CURRENT_SOURCE_DIR}/include/**.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/**.cpp
+list(
+  APPEND TARGET_FILES
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/main.cpp
 )
 
 add_executable(
-  ${SPORE_CODEGEN_EXAMPLES_JSON_TARGET}
-  ${SPORE_CODEGEN_EXAMPLES_JSON_FILES}
-)
-
-set_target_properties(
-  ${SPORE_CODEGEN_EXAMPLES_JSON_TARGET}
-  PROPERTIES
-  LINKER_LANGUAGE CXX
+  ${TARGET_NAME}
+  ${TARGET_FILES}
 )
 
 target_include_directories(
-  ${SPORE_CODEGEN_EXAMPLES_JSON_TARGET}
+  ${TARGET_NAME}
   PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/include
     ${CMAKE_CURRENT_SOURCE_DIR}/.codegen/include
@@ -30,7 +20,7 @@ target_include_directories(
 find_package(nlohmann_json CONFIG REQUIRED)
 
 target_link_libraries(
-  ${SPORE_CODEGEN_EXAMPLES_JSON_TARGET}
+  ${TARGET_NAME}
   PRIVATE
     nlohmann_json::nlohmann_json
 )
@@ -38,7 +28,12 @@ target_link_libraries(
 include(SporeCodegen)
 
 spore_codegen(
-  ${SPORE_CODEGEN_EXAMPLES_JSON_TARGET}
+  ${TARGET_NAME}
   CACHE .codegen/cache.yml
   DEBUG
+)
+
+set_target_properties(
+  ${TARGET_NAME} PROPERTIES
+  ADDITIONAL_CLEAN_FILES ${CMAKE_CURRENT_SOURCE_DIR}/.codegen
 )
